Add item condition to list and adjust image size
The updated code now includes the condition of the item in the list of attributes for all items. This provides the user with more detailed information about each auction item. The height of the item images have also been adjusted to 8rem to improve their display in the user interface.