Add auction location to item details
The code now scrapes the location of each auction and includes it as an attribute in the item details. This extra information will provide users with a better understanding of where the auction is taking place. The display in the user interface has been updated accordingly to show this additional field.